whatsapp_bot/requirements.txt
Flask==3.0.0
twilio==8.1.0
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
